requirements.txt
binwalk==2.4.3
pyshark==0.6
python-dotenv==1.2.1
python_netdiscover==0.3.0
python_nmap==0.7.1
python_owasp_zap_v2.4==0.0.14
python_owasp_zap_v2.4==0.0.20
Requests==2.32.5
simple_term_menu==1.6.6
